… OPTIMIZATION_SUMMARY.md
… lua/autorun/asc_performance_monitor.lua
… lua/autorun/hyperdrive_hyperspace.lua
… lua/autorun/hyperdrive_quantum.lua
… lua/autorun/hyperdrive_tactical_ai.lua
… lua/entities/asc_ship_core/init.lua